+++ client/ui/components/addons/AddonsList.jsx
@@ -7,10 +7,10 @@ import {
 } from 'material-ui/Table';
 import { Toolbar, ToolbarGroup } from 'material-ui/Toolbar';
 
-import FetchErrorList from './stateless/FetchErrorList';
-import LoadingIndicator from './stateless/LoadingIndicator';
-import containAddonGroupsList from '../containers/AddonGroupsListContainer';
-import goTo from '../utils/goTo';
+import FetchErrorList from './../error/FetchErrorList';
+import LoadingIndicator from '../indicators/LoadingIndicator';
+import containAddonsList from '../../containers/AddonsListContainer';
+import goTo from '../../utils/goTo';
 
 
 const style = {
@@ -22,23 +22,46 @@ const style = {
   },
 };
 
-class AddonGroupsList extends React.Component {
+class AddonsList extends React.Component {
   static propTypes = {
-    addonGroups: pt.array,
+    addons: pt.array,
     request: pt.object,
   };
 
-  renderRows() {
-    const { addonGroups } = this.props;
-
-    return addonGroups.map((addonGroup, index) =>
+  static renderRows(addons) {
+    return addons.map((addon, index) =>
       <TableRow key={index}>
         <TableRowColumn>
-          <Link to={`/addon_groups/${addonGroup.id}/`}>{addonGroup.browser_version}</Link>
+          <Link to={`/addons/${addon.id}/`}>{addon.name}</Link>
+        </TableRowColumn>
+        <TableRowColumn>
+          <Link to={`/addons/${addon.id}/`}>{addon.version}</Link>
         </TableRowColumn>
         <TableRowColumn className="align-right">
           <RaisedButton
-            onClick={() => goTo(`/addon_groups/${addonGroup.id}/`)}
+            onClick={() => goTo(`/addons/${addon.id}/`)}
+            label="Edit"
+            style={style.button}
+          />
+        </TableRowColumn>
+      </TableRow>
+    );
+  }
+
+  renderRows() {
+    const { addons } = this.props;
+
+    return addons.map((addon, index) =>
+      <TableRow key={index}>
+        <TableRowColumn>
+          <Link to={`/addons/${addon.id}/`}>{addon.name}</Link>
+        </TableRowColumn>
+        <TableRowColumn>
+          <Link to={`/addons/${addon.id}/`}>{addon.version}</Link>
+        </TableRowColumn>
+        <TableRowColumn className="align-right">
+          <RaisedButton
+            onClick={() => goTo(`/addons/${addon.id}/`)}
             label="Edit"
             style={style.button}
           />
@@ -71,8 +94,8 @@ class AddonGroupsList extends React.Component {
         <Toolbar style={style.toolbar}>
           <ToolbarGroup className="align-right" lastChild>
             <RaisedButton
-              label="Create New Addon Group"
-              onClick={() => goTo('/addon_groups/new/')}
+              label="Create New Addon"
+              onClick={() => goTo('/addons/new/')}
               primary
             />
           </ToolbarGroup>
@@ -80,7 +103,8 @@ class AddonGroupsList extends React.Component {
         <Table selectable={false}>
           <TableHeader displaySelectAll={false} adjustForCheckbox={false}>
             <TableRow>
-              <TableHeaderColumn>Browser Version</TableHeaderColumn>
+              <TableHeaderColumn>Name</TableHeaderColumn>
+              <TableHeaderColumn>Version</TableHeaderColumn>
               <TableHeaderColumn />
             </TableRow>
           </TableHeader>
@@ -93,4 +117,4 @@ class AddonGroupsList extends React.Component {
   }
 }
 
-export default containAddonGroupsList(AddonGroupsList);
+export default containAddonsList(AddonsList);
