+++ client/ui/components/pages/AddonGroupsListPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Paper from 'material-ui/Paper';
 
-import AddonGroupsList from '../AddonGroupsList';
+import AddonGroupsList from '../addon-groups/AddonGroupsList';
 import QueryAddonGroups from '../data/QueryAddonGroups';
