+++ client/ui/components/pages/AddonsListPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Paper from 'material-ui/Paper';
 
-import AddonsList from '../AddonsList';
+import AddonsList from '../addons/AddonsList';
 import QueryAddons from '../data/QueryAddons';
