+++ client/ui/components/pages/NotFoundPage.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import Paper from 'material-ui/Paper';
 
-import NotFound from '../NotFound';
+import NotFound from '../error/NotFound';
 
 
 function NotFoundPage() {
